Removed the redundant cast to the double type.
git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@12450 d02ae147-ee12-4a35-9452-91c89e8b2592

+++ src/com/surelogic/common/FileUtility.java
@@ -182,7 +182,7 @@ public final class FileUtility {
 		while (size > 1024 && labelIndex < 3) {
 			final long oldSize = size;
 			size = oldSize / 1024L;
-			digit = (int) (((oldSize % 1024L) / (double) 1024.0) * (double) 10.0);
+			digit = (int) (((oldSize % 1024L) / 1024.0) * 10.0);
 			labelIndex++;
 		}
 		if (labelIndex == 0 || digit == 0)
